% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{available_models}
\alias{available_models}
\title{list of available models in mafs package}
\usage{
available_models()
}
\value{
A character vector of the forecast models that can be used in this package.
}
\examples{
available_models()
}

