% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error_metrics}
\alias{error_metrics}
\title{list of available error metrics in mafs package}
\usage{
error_metrics()
}
\value{
A character vector of the error metrics that can be used in this package.
}
\description{
See forecast::accuracy() for more details.
}
\details{
There is an internal function in this package (removeTheil()) that
removes Theil'U metric from the output. This was necessary because for some
time series, forecast::accuracy() does not output the value for this
metric.
}
\examples{
error_metrics()
}

