% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gg_fit}
\alias{gg_fit}
\title{Graphical results of forecast models}
\usage{
gg_fit(x, test_size, model_name)
}
\arguments{
\item{x}{A ts object.}

\item{test_size}{Integer. The desired length of the test set object to be used
to train the model and compare the forecasted with the observed values.}

\item{model_name}{A string indicating the name of the forecast model.}
}
\value{
A ggplot object
}
\description{
Applys a selected forecast model to a time series and plot the fitted
series and the forecasted values along with the original series.
}
\examples{
gg_fit(AirPassengers, 12, "snaive")
}

