% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggplotFit}
\alias{ggplotFit}
\title{Plot a time series and its fitted and forecasted values}
\usage{
ggplotFit(x, test_size, horizon, model_name)
}
\arguments{
\item{x}{A ts object.}

\item{test_size}{The desired length of the test set object to be used
to measure the accuracy of the forecast models.}

\item{horizon}{The forecast horizon length}

\item{Name}{of the forecast model. Please run available_models() to see the full list.}
}
\value{
A ggplot object
}
\description{
Uses ggplot to plot a given time series as well as its fitted and
forecasted values obtained with one of the forecast functions available
in mafs package.
}
\details{
TODO
}
\examples{
ggplotarFit(AirPassengers, 6, 6, "ets")
}

