% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_forecast}
\alias{select_forecast}
\title{Selects best forecast model}
\usage{
select_forecast(x, test_size, horizon, error, dont_apply = "",
  verbose = FALSE)
}
\arguments{
\item{x}{A ts object.}

\item{test_size}{The desired length of the test set object to be used
to measure the accuracy of the forecast models.}

\item{horizon}{The forecast horizon length}

\item{error}{The accuracy metric to be used to select the best forecast
model from apply_all_models(). See error_metrics() for the available metrics.}

\item{dont_apply}{Character vector. Choose one or more models that will not
be used on select_forecast().}

\item{verbose}{logical. Set TRUE if you want mafs to tell you what models are running.}
}
\value{
A list of three objects:
}
\description{
select_forecast is the main function of this package. It uses
  apply_all_models() and other internal functions of this package to generate
  generate multiple forecasts for the same time series object.
}
\details{
TODO
}
\section{df_models}{

 A data.frame with the accuracy metrics of all models applied to x
}

\section{best_forecast}{

 A forecast object created by applying the best forecast method to x
}

\section{df_comparison}{

 A dataframe showing both the forecasted and the observed test set
}

\examples{
\dontrun{
select_forecast(austres, 6, 12, "MAPE")
}
}
