% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{splitTrainTest}
\alias{splitTrainTest}
\title{Split Time Series Data into Training and Testing Sets}
\usage{
splitTrainTest(dat, numTrain = length(dat) - 10)
}
\arguments{
\item{dat}{A time series object (of class `ts`).}

\item{numTrain}{An integer specifying the number of observations to include in the training set.
Defaults to `length(dat) - 10`.}
}
\value{
A list containing:
	\item{train}{A time series object containing the training data.}
	\item{test}{A time series object containing the testing data.}
}
\description{
This function splits a time series object into training and testing sets.
}
\examples{
ts_data <- ts(1:100, start = c(2000, 1), frequency = 12)
split_data <- splitTrainTest(ts_data, numTrain = 80)
train_set <- split_data$train
test_set <- split_data$test

}
